Update tmux, zsh and brew
This commit adjust a few visual properties on tmux, zsh and the songs
script for macos.

Add llvm as a core dependency into the Brewfile.

Also replace the base github uri of the modules from ssh to https, as
such it is not dependent of setting up ssh first and we can have all
core vim modules installed at once.