Add lua-language-server and textDocument/rename custom handler
This commit adds the setup for the lua-language-server, configured for
neovim development (global vim) and add a new custom handler for method
textDocument/rename in which we send all replaced spots to the quickfix
list in order to better investigate the changes.

Also a few minor adjustments to jdtls, lspconfig, telescope and treesitter
configurations.